Refactor DepositPage component to integrate form and verification steps, replacing DepositPageController and DepositChainSpecificProvider with a streamlined implementation. Update DepositFeeValue and DepositFiatFeeValue to utilize new query hooks for improved data handling and loading states.
