+++ core/ui/vault/deposit/fee/DepositFiatFeeValue.tsx
@@ -7,9 +7,12 @@ import { useFiatCurrency } from '@core/ui/storage/fiatCurrency'
 import { useCurrentVaultCoin } from '@core/ui/vault/state/currentVaultCoins'
 import { Spinner } from '@lib/ui/loaders/Spinner'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
+import { useTransformQueriesData } from '@lib/ui/query/hooks/useTransformQueriesData'
 import { formatAmount } from '@lib/utils/formatAmount'
+import { useCallback } from 'react'
+import { useTranslation } from 'react-i18next'
 
-import { useDepositChainSpecific } from './DepositChainSpecificProvider'
+import { useDepositChainSpecificQuery } from '../queries/useDepositChainSpecificQuery'
 
 export const DepositFiatFeeValue = () => {
   const [{ coin: coinKey }] = useCoreViewState<'deposit'>()
@@ -20,24 +23,35 @@ export const DepositFiatFeeValue = () => {
 
   const fiatCurrency = useFiatCurrency()
 
-  const chainSpecific = useDepositChainSpecific()
+  const chainSpecificQuery = useDepositChainSpecificQuery()
 
   const { decimals } = chainFeeCoin[coinKey.chain]
 
-  const fee = getFeeAmount(chainSpecific)
+  const query = useTransformQueriesData(
+    {
+      price: priceQuery,
+      chainSpecific: chainSpecificQuery,
+    },
+    useCallback(
+      ({ price, chainSpecific }) => {
+        const fee = getFeeAmount(chainSpecific)
 
-  const feeAmount = fromChainAmount(fee, decimals)
+        const feeAmount = fromChainAmount(fee, decimals)
+
+        return formatAmount(feeAmount * price, fiatCurrency)
+      },
+      [decimals, fiatCurrency]
+    )
+  )
+
+  const { t } = useTranslation()
 
   return (
     <MatchQuery
-      value={priceQuery}
+      value={query}
       pending={() => <Spinner />}
-      error={() => null}
-      success={price => {
-        const formattedAmount = formatAmount(feeAmount * price, fiatCurrency)
-
-        return formattedAmount
-      }}
+      error={() => t('failed_to_load')}
+      success={result => <>{result}</>}
     />
   )
 }
